requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pytz==2022.7.1
sqlparse==0.4.3
wrapt==1.11.2
django-filter==22.1
django-cors-headers==3.13.0

cryptography~=39.0.0
pip~=22.2.2
attrs~=22.1.0
distro~=1.7.0
wheel~=0.37.1
Pillow~=9.0.1
Jinja2~=3.1.2
keyring~=23.5.0
platformdirs~=2.5.2
colorama~=0.4.4
click~=8.0.3
psycopg2~=2.9.5
PyYAML~=5.4.1
requests~=2.24.0
Markdown~=3.4.1
setuptools~=59.6.0
zipp~=3.11.0
Faker~=16.6.0